+++ src/App.js
@@ -16,9 +16,21 @@ class App extends Component {
     
     console.log(`Found token ${token} for username ${username}.`);
     this.state = {
+      /**
+       * @param {string|null}
+       */
       token: (token === undefined || token === "") ? null : token,
+      /**
+       * @param {string|null}
+       */
       username: username === undefined ? null : username,
+      /**
+       * @param {[object]}
+       */
       lists: [],
+      /**
+       * @param {object|null}
+       */
       selected_list: null,
     }
 
@@ -115,14 +127,41 @@ class App extends Component {
     return json;
   }
 
-  async addChore(name, checklist_pk) {
+  async addTask(name, frequency, checklist_pk) {
     console.log(`Adding ${name} to checklist ${checklist_pk}...`);
-    this.API.post(`chores/api/checklist/${checklist_pk}/add_chore`, {}, {name: name});
-    await this.selectList(this.state.selected_list.id);
+    let response = await this.API.post(
+      `chores/api/checklist/${checklist_pk}/add_chore`,
+      {},
+      {name: name, frequency: frequency}
+    );
+    if (response.status === 200) {
+      let newItem = await response.json();
+      console.log('Adding', newItem, 'to the local list.');
+      let newList = {
+        ...this.state.selected_list,
+        items: [...this.state.selected_list.items, newItem]
+      };
+      this.setState({selected_list: newList});
+      console.log('Added.');
+    }
+    console.log('An error has occured.');
+    return response;
   }
 
-  updateChore(pk, name, checklist) {
-  
+  async updateChore(pk, name, frequency) {
+    console.log(`Updating chore ${pk} to name ${name} and freq. of ${frequency} days...`);
+    let response = await this.API.post(
+      `chores/api/chore/${pk}/update`,
+      {},
+      {
+        name: name,
+        frequency: frequency
+      }
+      );
+    if (response.status === 200) {
+      await this.selectList(this.state.selected_list.id);
+      return response.json();
+    }
   }
 
   async logChore(pk, note="", date=null) {
@@ -136,8 +175,22 @@ class App extends Component {
     return await response.json();
   }
 
-  deleteChore(pk) {
+  async deleteTask(pk) {
+    console.log(`Deleting task ${pk}...`);
+    let response = await this.API.get(`chores/api/chore/${pk}/delete`);
+    if (response.status === 200) {
+      console.log(`Deleting task ${pk} from the local list.`);
+      let newList = {
+        ...this.state.selected_list,
+        items: [...this.state.selected_list.items.filter(x => x.id !== pk)]
+      };
 
+      let json = await response.json();
+      this.setState({selected_list: newList});
+      console.log(`Deleted.`);
+      return json;
+    }
+    console.log(`An error occured during deleting task ${pk}.`);
   }
   
   async deleteLog(pk) {
@@ -185,10 +238,11 @@ class App extends Component {
             <MainView
               updateName={(name) => this.updateList(this.state.selected_list.id, name)}
               deleteList={() => this.deleteList(this.state.selected_list.id)}
-              addTask={(name) => this.addChore(name, this.state.selected_list.id)}
+              addTask={(name, frequency) => this.addTask(name, frequency, this.state.selected_list.id)}
               getChore={(pk) => this.getChore(pk)}
+              updateTask={(pk, name, frequency) => this.updateChore(pk, name, frequency)}
               logChore={(pk, note, dtg) => this.logChore(pk, note, dtg)}
-              deleteChore={pk => this.deleteChore(pk)}
+              deleteTask={pk => this.deleteTask(pk)}
               deleteLog={pk => this.deleteLog(pk)}
               list={this.state.selected_list}/>
           </> }